updated to clean up code based on PR comments
