+++ lib/ditto_flutter_tools.dart
@@ -14,4 +14,4 @@ export "src/sync_status_view.dart" show SyncStatusView;
 export "src/sync_status_helper.dart" show SyncStatusHelper, SyncStatus;
 export "src/peer_list.dart" show PeerListView;
 export "src/system_settings_view.dart" show SystemSettingsView;
-export "src/query_editor_view.dart" show QueryEditorView;
+export "src/query_editor/query_editor_view.dart" show QueryEditorView;
